Refactor - Split Predicate methods
