+++ src/DynamicQueryable/DynamicQueryable.cs
@@ -4,38 +4,10 @@ using Jokenizer.Net;
 
 namespace System.Linq.Dynamic {
 
-    public static class DynamicQueryable {
+    public static partial class DynamicQueryable {
 
         public static IQueryable<T> As<T>(this IQueryable source) {
             return (IQueryable<T>)source;
-        }
-
-        public static IQueryable<T> Where<T>(this IQueryable<T> source, string predicate, params object[] values) {
-            return Where<T>(source, predicate, null, values);
-        }
-
-        public static IQueryable<T> Where<T>(this IQueryable<T> source, string predicate, IDictionary<string, object> variables, params object[] values) {
-            return (IQueryable<T>)Where((IQueryable)source, predicate, variables, values);
-        }
-
-        public static IQueryable Where(this IQueryable source, string predicate, params object[] parameters) {
-            return Where(source, predicate, null, parameters);
-        }
-
-        public static IQueryable Where(IQueryable source, string predicate, IDictionary<string, object> variables, params object[] parameters) {
-            if (source == null) throw new ArgumentNullException(nameof(source));
-            if (predicate == null) throw new ArgumentNullException(nameof(predicate));
-
-            var types = new[] { source.ElementType };
-            var lambda = Evaluator.ToLambda(predicate, types, variables, parameters);
-
-            return source.Provider.CreateQuery(
-                Expression.Call(typeof(Queryable),
-                "Where",
-                types,
-                source.Expression,
-                Expression.Quote(lambda))
-            );
         }
 
         public static IQueryable Select(this IQueryable source, string selector, params object[] values) {
@@ -225,66 +197,6 @@ namespace System.Linq.Dynamic {
                     new[] { source.ElementType, keyLambda.Body.Type },
                     source.Expression,
                     Expression.Quote(keyLambda)
-                )
-            );
-        }
-
-        public static IQueryable<T> SkipWhile<T>(this IQueryable<T> source, string predicate, params object[] values) {
-            return SkipWhile(source, predicate, null, values);
-        }
-
-        public static IQueryable<T> SkipWhile<T>(this IQueryable<T> source, string predicate, IDictionary<string, object> variables, params object[] values) {
-            return (IQueryable<T>)SkipWhile((IQueryable)source, predicate, variables, values);
-        }
-
-        public static IQueryable SkipWhile(this IQueryable source, string predicate, params object[] values) {
-            return SkipWhile(source, predicate, null, values);
-        }
-
-        public static IQueryable SkipWhile(this IQueryable source, string predicate, IDictionary<string, object> variables, params object[] values) {
-            if (source == null) throw new ArgumentNullException(nameof(source));
-            if (predicate == null) throw new ArgumentNullException(nameof(predicate));
-
-            var types = new[] { source.ElementType };
-            var lambda = Evaluator.ToLambda(predicate, types, variables, values);
-
-            return source.Provider.CreateQuery(
-                Expression.Call(
-                    typeof(Queryable), 
-                    "SkipWhile",
-                    new Type[] { source.ElementType },
-                    source.Expression, 
-                    Expression.Quote(lambda)
-                )
-            );
-        }
-
-        public static IQueryable<T> TakeWhile<T>(this IQueryable<T> source, string predicate, params object[] values) {
-            return TakeWhile(source, predicate, null, values);
-        }
-
-        public static IQueryable<T> TakeWhile<T>(this IQueryable<T> source, string predicate, IDictionary<string, object> variables, params object[] values) {
-            return (IQueryable<T>)TakeWhile((IQueryable)source, predicate, variables, values);
-        }
-
-        public static IQueryable TakeWhile(this IQueryable source, string predicate, params object[] values) {
-            return TakeWhile(source, predicate, null, values);
-        }
-
-        public static IQueryable TakeWhile(this IQueryable source, string predicate, IDictionary<string, object> variables, params object[] values) {
-            if (source == null) throw new ArgumentNullException(nameof(source));
-            if (predicate == null) throw new ArgumentNullException(nameof(predicate));
-
-            var types = new[] { source.ElementType };
-            var lambda = Evaluator.ToLambda(predicate, types, variables, values);
-
-            return source.Provider.CreateQuery(
-                Expression.Call(
-                    typeof(Queryable), 
-                    "TakeWhile",
-                    new Type[] { source.ElementType },
-                    source.Expression, 
-                    Expression.Quote(lambda)
                 )
             );
         }
